🤪Intercept test framework invocation
Intercepting `test_main_static` via patching the function code so we
get references to all of the built-in tests without the need for
our custom `#[test]` attribute and custom entry point `datatest::harness!`.

Brings behavior on stable Rust to the same level as on nightly at the
cost of relying on a terrible hack that might explode at any point in
time.