Fix typo in comment about the handling of #[files(...)] syntax.
In the second paragraph which mentiones the trampoline function,
there's a mention about trait magic to make functions work for both
arguments which take a slice and owned arguments.

For the slice the comment gives as examples '&str' and '&[u8'.
The second example had a minor typo: the closing square bracket
was missing.